Move generators/type.js to generate.js
