+++ lib/generate.js
