+++ lib/index.js
@@ -8,7 +8,7 @@ var parse = require('./parse')
 /**
  * Generators
  */
-var Type = require('./generators/type')
+var Type = require('./generate')
 
 /**
  * Export `Create`
